clean code 清理
# message auto-generated for no-merge-commit merge:
merge clean_code into master

clean code 清理

Created-by: insightDeveloper

Match-id-2a140422a249671fd9071b3dff4d8d4e288efa4c

+++ server/src/modules/global/database/SystemMemoryDatabase.cpp
@@ -295,8 +295,8 @@ std::map<int64_t, std::vector<std::shared_ptr<ParseFileInfo>>> SystemMemoryDatab
 
     while (resultSet->Next()) {
         auto info = std::make_shared<ParseFileInfo>();
-        info->id = resultSet->GetUint64("id");
-        info->projectExplorerId = resultSet->GetUint64("projectExplorerId");
+        info->id = static_cast<int64_t>(resultSet->GetUint64("id"));
+        info->projectExplorerId = static_cast<int64_t>(resultSet->GetUint64("projectExplorerId"));
         info->parseFilePath = resultSet->GetString("parseFilePath");
         info->fileId = resultSet->GetString("fileId");
         info->subId = resultSet->GetString("subId");
