+++ server/src/modules/summary/core/ExpertHotspotManager.cpp
@@ -75,7 +75,8 @@ bool ExpertHotspotManager::MergeAndSaveModelInfo(const std::map<std::string, Mod
     if (!hotspotModelInfo.empty()) {
         // 计算总层数，从以下两个数据中取大的数据：1.导入数据的moe层数+已配置的稠密层数；2.已配置的总层数
         uint64_t totalModelLayerCeil = NumberSafe::Add(hotspot.moeLayer, curModelInfo.denseLayerList.size());
-        int totalModelLayer = NumberUtil::CeilingClamp(totalModelLayerCeil, static_cast<uint64_t>(INT_MAX));
+        int totalModelLayer = static_cast<int>(NumberUtil::CeilingClamp(totalModelLayerCeil,
+                                                                        static_cast<uint64_t>(INT_MAX)));
         finalModelInfo.modelLayer = std::max(totalModelLayer, finalModelInfo.modelLayer);
         finalModelInfo.rankNumber = hotspot.rankNumber;
         finalModelInfo.moeLayer = hotspot.moeLayer;
@@ -84,7 +85,8 @@ bool ExpertHotspotManager::MergeAndSaveModelInfo(const std::map<std::string, Mod
 
     if (!deploymentModelInfo.empty()) {
         uint64_t totalModelLayerCeil = NumberSafe::Add(deployment.moeLayer, curModelInfo.denseLayerList.size());
-        int totalModelLayer = NumberUtil::CeilingClamp(totalModelLayerCeil, static_cast<uint64_t>(INT_MAX));
+        int totalModelLayer = static_cast<int>(NumberUtil::CeilingClamp(totalModelLayerCeil,
+                                                                        static_cast<uint64_t>(INT_MAX)));
         finalModelInfo.modelLayer = std::max(totalModelLayer, finalModelInfo.modelLayer);
         finalModelInfo.rankNumber = deployment.rankNumber;
         finalModelInfo.moeLayer = deployment.moeLayer;
